+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -278,7 +278,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K6CEZNTCVYB7FMDFC7ES59MY"),
+  _id: z.string().default("01K6D0K7PW08VRDAN1RVXGA5X7"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -322,7 +322,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01K6CEZNTCVYB7FMDFC7ES59MY"),
+  id: z.string().default("01K6D0K7PW08VRDAN1RVXGA5X7"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,