+++ packages/orq-rc/RELEASES.md
@@ -3959,3 +3959,13 @@ Based on:
 - [typescript v3.13.0-rc.48] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.48 - packages/orq-rc
+
+## 2025-09-30 09:54:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.5 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.49] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.49 - packages/orq-rc